C - Binary Trees. Task 12. Leaves

+++ 9-binary_tree_height.c
@@ -10,7 +10,7 @@
  */
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int height_l = 0, height_r = 0;
+	size_t height_l = 0, height_r = 0;
 
 	if ((tree == NULL) || ((tree->left == NULL) && (tree->right == NULL)))
 		return (0);
